requirements.dev.txt
flake8==7.1.1
pytest>=7.4.0
pytest-django>=4.5.2
pytest-cov>=4.1.0
coverage>=7.2.7